Refactor functions to use common code
