+++ Makefile
@@ -8,7 +8,11 @@ input := $(shell mktemp -u)
 default:
 	docker build -t quay.io/mojodna/osm-pds-pipelines .
 
-deploy: project.json
+deploy: project.json node_modules/.bin/interp
+	for f in functions/* ; do \
+		cd $$f && npm install; cd -; \
+		interp < $$f/function.json.hbs > $$f/function.json; \
+	done
 	apex deploy
 
 install: project.json
