+++ functions/mirror/function.json.hbs
@@ -1,0 +1,8 @@
+{
+  "description": "Mirror OSM planet dumps to S3",
+  "runtime": "nodejs4.3",
+  "environment": {
+    "BATCH_JOB_QUEUE": "osm-transcoding",
+    "S3_BUCKET": "osm-pds"
+  }
+}
