+++ sample.env
@@ -4,3 +4,4 @@ AWS_ACCOUNT_ID=<redacted>
 AWS_REGION=us-east-1
 BATCH_JOB_QUEUE=osm-transcoding
 S3_BUCKET=osm-pds
+S3_CHECKPOINT_BUCKET=osm-pds-status
